# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DealPilot Local API
  version: 1.0.0
  description: |
    DealPilot 本地优先架构的后端 API 契约。
    所有端点前缀 /api/v1，除 /health 外全部需要 Bearer Token 认证。
    Agent 只监听 127.0.0.1，禁止 0.0.0.0。
    列表响应使用游标分页（cursor + next_cursor）。
    写操作支持 Idempotency-Key 请求头实现幂等重试。
  contact:
    name: DealPilot
servers:
  - url: http://127.0.0.1:{port}
    description: 本地 Agent 回环地址
...
paths:
...
  /api/v1/customers/{id}/social-accounts:
    get:
      tags: [customers]
      summary: 社媒账号列表
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 社媒账号列表
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SocialAccount' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [customers]
      summary: 添加社媒账号
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SocialAccountCreate' }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SocialAccount' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/social-accounts/{id}:
    delete:
      tags: [customers]
      summary: 删除社媒账号
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      responses:
        '204':
          description: 删除成功
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string }
...
    IdempotencyKeyParam:
      name: Idempotency-Key
      in: header
      description: 幂等键，重复请求返回首次结果
      schema: { type: string }
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: 短期访问令牌，工作台来自 sessionStorage，插件来自 Native Messaging Bootstrap。
  responses:
    BadRequest:
      description: 请求参数或格式错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: 未认证或令牌无效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Origin 校验失败或权限不足
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: 资源冲突或状态非法
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    InternalError:
      description: 服务器内部错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: 错误码
        message:
          type: string
          description: 错误信息
        details:
          type: object
          description: 附加详情
        request_id:
          type: string
          description: 请求追踪 ID
      required: [code, message, request_id]
...
    SocialAccount:
      type: object
      properties:
        id: { type: string }
        customer_id: { type: string }
        contact_id: { type: string, nullable: true }
        platform: { type: string }
        raw_identifier: { type: string }
        normalized_identifier: { type: string }
        manually_bound: { type: boolean }
        created_at: { type: string, format: date-time }
      required: [id, customer_id, platform, raw_identifier, normalized_identifier, manually_bound, created_at]
    SocialAccountCreate:
      type: object
      properties:
        platform: { type: string }
        raw_identifier: { type: string }
        contact_id: { type: string }
      required: [platform, raw_identifier]
...
security:
  - BearerAuth: []